server/requirements.txt
beautifulsoup4==4.12.0
EbookLib==0.17.1
epub2txt==0.1.1
fastapi==0.95.0
langchain==0.0.144
llama_index==0.5.17.post1
pydantic==1.10.7
PyPDF2==3.0.1
requests==2.28.2
starlette==0.26.1
uvicorn==0.21.1
docx2txt==0.8
html2text==2020.1.16
ipython==8.12.2
python_multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==21.3.0